LRAC-14780 analytics-machine-learning-impl Move the RecommendationField to the search package
