+++ modules/apps/analytics/analytics-machine-learning-impl/src/main/java/com/liferay/analytics/machine/learning/internal/recommendation/search/RecommendationField.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.analytics.machine.learning.internal.recommendation.constants;
+package com.liferay.analytics.machine.learning.internal.recommendation.search;
 
 /**
  * @author Riccardo Ferrari
